Fix usage of NULL #define
The NULL #define normally specifies a null pointer, i.e it is
typically defined as something like ((void *)0).  In EmuTOS,
it was defined as 0, and sometimes used as zero, sometimes as
a nul character ('\0'), and sometimes as a null pointer.
This update changes the #define to the standard one, and
changes any usage of NULL as an integer or character to 0
or '\0' as appropriate.