Allocate screen objects dynamically
The previous code had a small, fixed number of such objects. This
meant that in any screen resolution higher than ST low, it was
possible to open a disk or folder and be unable to display all the
contents. The number of screen objects is now adjusted according
to screen resolution to prevent this happening.