Always display untranslated text in original form
When building a multi-language ROM, a string in a .po file that
did not have a translation (i.e. whose msgid was empty) was always
displayed in its original form (i.e. msgstr). However, for single
language ROMs, the same string was displayed as an empty string.
The behaviour is now fixed to be the same as for multi-language.