Change -More- message for EmuCON
The -More- string was used by both the desktop and EmuCON when
displaying a text file longer than a page, which caused the
translation to be shared. To allow EmuCON to remain in English
while translating the desktop (requested by some translators),
slightly different messages are now used. Thanks to Eero
Tamminen for pointing this out.