Ensure desktop menu bar fits within screen
In ST Low, translations could cause the menu bar to be truncated
when displayed because it was too wide and/or it started too far
to the right.  This fix moves the starting point to the left if
required.  If the menu bar would still be truncated during display,
the last heading is explicitly truncated to avoid display artifacts.