Fix obscure bug in raster operations
The raster operations vro_cpyfm(), vrt_cpyfm() and lineA copy raster
are passed the coordinates of each of the source and destination
rectangles as 4 values: (x1,y1),(x2,y2) where (x1,y1) is the upper
left corner and (x2,y2) is the lower right.  There is an undocumented
feature in TOS that is also replicated in EmuTOS: if x1>x2, x1 & x2
are swapped; if y1>y2, y1 & y2 are swapped.
However, there was a casting bug in EmuTOS (introduced by commit
421d27 in 2004) that caused this not to be done for the destination
rectangle.  This bug caused the problem with the melt-down effect in
Mario Becroft's Pacman game.