Mix .bss and COMMON sections
This trick allows the linker to fully respect the order of the object
files specified on the command line.
Previously, .bss and COMMON were 2 separate blocks in the BSS segment.

As a result, all the BDOS variables are located after the BIOS ones,
and so on. Critical variables can be kept as low as possible in the
RAM, to protect them from rude user programs which may overwrite
low memory areas without caring of anything.