Put global C variables into .bss section instead of COMMON
This is achieved with the -fno-common GCC option.

The goal is to simplify the allocation of the BSS contents,
in order to make emutos.map more readable.

Since the COMMON section is now empty, the *(.bss COMMON) hack
in emutos.ld to mix .bss and COMMON contents is now pointless.

Furthermore, -fno-common prevents to define the same global variable
in multiple source files. That would be a bad idea, anyway.