New notion of "first boot" instead of "cold boot"
A first boot is the first time this EmuTOS instance runs.
A cold boot from ROM is always a first boot.
A reset is never a first boot.

When EmuTOS runs from RAM for the first time, it is also a first boot,
while it is not a cold boot.