Add _5MS cookie on non-Atari hardware
On Atari hardware, the 200 Hz system timer is implemented with the MFP
Timer C. Its interrupt vector is located at address 0x114.

On non-Atari hardware, the 200 Hz system timer is implemented differently,
so it uses a different interrupt vector. We define the _5MS cookie
with the address of the 200 Hz system timer interrupt vector so FreeMiNT
can hook it.