Fix VgetRGB()/vq_color() for the Falcon again
For VgetRGB()/vq_color() on a Falcon, we must get the current values
of the palette registers from the shadow registers rather than the
real hardware registers, since the next VBL interrupt may update the
real registers from the shadow ones.

We were already doing this for the videl registers; this fix does it
for the STe-compatible registers too.