Fix os_conf value and usage in multilanguage ROMs
Previously, os_conf was erroneously set to 0xfe in multilanguage ROMs.
Issue reported by Dima Sobolev.

If os_conf is set to 0xff, the default language and PAL/NTSC mode are
inferred from the COUNTRY Makefile variable. Otherwise, they are taken
from the os_conf value itself.

The os_conf value can be patched in the ROM header to select a different
default language and PAL/NTSC mode when the NVRAM is not present.

In any case, the display and keyboard languages will be overridden by the
NVRAM settings if NVRAM is present and valid.