Do not change mouse pointer to arrow prematurely
When doing File|Info on a disk, the mouse pointer was changed to an
arrow before the total free space was determined.  Since this can
be a lengthy process, it could be a long time between the mouse
changing to an arrow and the info dialog appearing.  This could
look like a system crash, at least briefly.  Now the pointer is
changed after the free space is determined.