Fix bug when File Selector accesses empty drive
When accessing an empty drive with the File Selector, the critical
error handler dialog is displayed.  Previously, selecting Cancel
would cause the invalid path to be displayed (with no files shown).
Now it causes an appropriate form_alert() to be issued, followed by
a redisplay of the previous path.  This is the same as Atari TOS.