Use strcpy() instead of strcat()
This fixes compilation with GCC 6.1.0.

Since p is at the end of ptail, strcpy() is equivalent to strcat() but
simpler in this case. So gcc replaces the strcat() call with strcpy().
But as gcc is unaware of USE_STATIC_INLINES, this causes an undefined
reference to _strcpy. Explicltly calling strcpy() avoids the problem.