Display hourglass immediately when opening folder
Previously, the folder was selected in the BDOS before calling
do_diropen().  When opening a floppy disk from the desktop on real
hardware, this caused the mouse pointer to remain an arrow for a
while, even though the system was busy.  Now the selection takes
place after the pointer has been changed to an hourglass.