Split TT-RAM detection and Alt-RAM initialization
This fixes a regression introduced in commit def7047.
The _FRB cookie was not properly installed, because
fill_cookie_jar() is called before altram_init().

Now the TT-RAM is detected in machine_detect(), before
fill_cookie_jar(). So the _FRB cookie is properly installed.

This can be tested with:
- TT hardware
- FastRAM
- ACSI disk (needs _FRB cookie)
Then run some program from disk.