Fix regression caused by commit 6511a65
The above commit changed the freeit() function so that it updated
the link to the next MD on the allocated list (previously, freeit()
expected the caller to do that).  This change inadvertently caused
free_all_owned() to update the pointer to the allocated list if the
first block on that list was to be freed, and bad things happened
after that.