Revert commit 410326
Commit 410326 was an initial attempt to address the "40-folder
problem" that EmuTOS (and the original Atari TOS 1.0) inherited
from DRDOS.  It ameliorated the situation, but did not fix the
root cause, which was finally fixed by commit 49d837.  With that
fix, commit 410326 now serves no useful purpose; reverting it
saves about 270 bytes in the 192K ROMs.