Fix bug in desktop window save
The status of a desktop window that was opened via an alt key was
not saved properly: specifically, the field corresponding to the
obid_save field was incorrect. This applied to both the EMUDESK.INF
file created via "Save desktop" and the information saved internally
when the desktop exited in order to run a program. As a result, the
window was not restored when the desktop restarted. This bug was
introduced by commit 44791b.