Fix bug in updating date when month rolls over
This affected the internal variable 'current_date'.  When the day in
'current_date' was 31, and the 'current_time' variable was updated
from 23:59:58 to 00:00:00, the day in 'current_date' was set to 0
rather than 1.  The same bug also caused the next date/time after
1997/12/31 23:59:58 to be 1997/13/00 00:00:00.
This bug was introduced by commit ef4f790.