Restore 'out of OS memory' recovery code
This reverts commit 6132d9c.

This code was originally added by commit 410326d as an attempt to
fix the "40-folder problem" that EmuTOS (and the original Atari TOS
1.0) inherited from DRDOS.  It ameliorated the situation, but did not
fix the most common cause, which was finally fixed by commit 49d837.
At that point, commit 410326d was thought to be unnecessary, and was
reverted by commit 6132d9c.  However, recent testing shows that the
original code is useful in some situations, so it is restored.