Simplify conditional compilation for ColdFire
In several places, EmuTOS saves/restores two registers on the stack.
This was done via movem.l for 680x0s, and 2*move.l for ColdFire.  It
is now done via 2*move.l for all systems, since this takes the same
amount of ROM space as a movem.l, and is faster on both 68000 and
68030 processors.