Falcon: Fix invalid sysvars erasure on reset
Consequently, boot drive was forced to A: on reset, preventing C:\AUTO
programs to run.

This bug was introduced in commit 42e4d35 when first_boot was moved to a
single bit in meminit_flags. But it only became visible in commit 6b50c0c
when a second bit, MEMINIT_BIT_FALCON_MMU, was introduced.
A test from full byte to single bit had been missed, forcing invalid
sysvars erasure on reset, on machines with Falcon MMU.

Thanks to Ingo Uhlemann, Roger Burrows and Christian Zietz for hunting
this bug.