Fix Rwabs() with "ignore mediachange" bit set
First time access to a floppy disk via Rwabs() failed with error
ESECNF, even if the "ignore media change status" bit was set in
the Rwabs() call.  The error continued until a Getbpb() was done.
This fix makes EmuTOS behave the same way as Atari TOS in this
situation.  Thanks to Thorsten Otto for discovering the problem.