Fix bug with desktop move/copy
When moving or copying a file/folder, if the target name already
exists, the desktop should allow the user to change it.  This worked
when the source and target were on different drives, but did not
work when the move/copy was within the same drive.
Thanks to Eero Tamminen for finding this.