Simplify cross-directory rename of a directory
Previously, if a directory was being renamed and it had an active
DND, we attempted to update the DND to correspond to the new
directory location.  This was complex and easy to get wrong,
resulting in a crash.  Now we just free the DND: it will be
recreated later if required.