Ensure floppy formatter triggers mediachange
Formatting a floppy should force mediachange but did not.  This
could cause several problems; for example, if a diskette was
formatted as HD, then reformatted as DD without removing the
diskette, the free space reported was the free space after HD
formatting.  This bug was introduced by commit 17a77ae.
Thanks to Thorsten Otto for reporting this.