Simplify offset calc when moving desk icons
The offsets in question are those from the icon origin to the point
where the mouse pointer has grabbed the icon.  This value is used
when determining where to drop the icon(s).

This change saves about 60 bytes in the 192K ROMs, and prepares for
an upcoming fix to the 'install desktop shortcut' code.