Fix bug with handling of installed window icons
If an icon was associated with a program file via "Install icon",
that file was subsequently treated as a non-executable file, and so
could not be launched by the desktop.  This problem has existed since
"Install icon" was introduced by commit b88d643 (in release 0.9.7).
Problem reported by Dima Sobolev.