Avoid translating TEDINFO text twice
The text in the te_ptmplt field of a TEDINFO object was translated
in xlate_obj_array(), and then the translation was translated again
in xlate_fix_tedinfo().  This did not cause any problems in practice,
since the second translation failed leaving the first translation OK,
but it consumes time & space, and in theory it could have caused a
mistranslation as well.  Thanks to Thorsten Otto for spotting this.