Remove check for valid tree in menu_tnormal()
Previously, under EmuTOS, menu_tnormal() checked that the specified
tree was the tree for the current menu bar; if not, the call silently
failed.  This check does not occur under Atari TOS or other AESs.
The game "Patience" was inadvertently relying on this to clear the
SELECTED bit of its dialog exit buttons via menu_tnormal(), so the
bit was not cleared under EmuTOS, producing buggy behaviour.
Removing this check fixes this and reduces EmuTOS code size too.