Implement KINFO() / KDEBUG() using NULL_FUNCTION()
The previous implementation replaced them with nothing in some cases.
This was correct, as the semicolon immediately following the macro was
acting as null instruction. This was correct even when using if / else
statements without braces. But it could have produced unexpected results
if someone had forgot that ending semicolon.

This new implementation is more robust. Instead of introducing a bug,
it will not compile if someone forgets the semicolon.