Fix potential desktop crash when opening folder
When a folder was opened, the object that triggered the open was
always redrawn.  When the folder was opened by clicking on it in an
existing window, this could cause a crash because the FNODE for the
folder had already been freed.  See the comments for do_wopen() for
more details.
Many thanks to Thorsten Otto for the bug report and analysis.