Always issue alert if no windows are available
When a desktop folder shortcut is clicked on, a new window is opened;
this requires allocating both a WNODE and a PNODE.  When the maximum
number of windows is reached, both allocations will fail.
Previously the PNODE was allocated first, and if it could not be
allocated, the window creation failed silently.  Now the WNODE is
allocated first, and failure will cause an alert to be issued.