Allow configuration of window/desktop backgrounds
This allows the background fill pattern & colour to be set for the
desktop and windows, via a desktop dialog.  The equivalent feature
is available in Atari TOS2/3/4.  This feature is disabled by default
for the 192K ROMs, since it is quite expensive in ROM space.
This update was inspired by Thorsten Otto.