Fix window icon display with non-white background
When files were displayed as icons in a window with a non-white
background, the icon mask was not drawn.  The resulting display was
unlike the desktop display, unlike Atari TOS, and not easy to read.
This problem was introduced when backgrounds became configurable,
in commit 798e403.  Now, if backgrounds are configurable, the mask
is always drawn, and the display is always 'normal'.