Allow desktop window file mask to be specified
By default, the file mask for a window is *.*, i.e. all files in the
selected folder are displayed.  The new menu item 'Set file mask...'
allows you to specify an alternate filemask to restrict the display,
just like Atari TOS 2/3/4.  Note that folders are always displayed,
irrespective of file mask.
This feature is disabled by default in the 192K ROMs.