Add blitter menu item to desktop
The Blitter menu item now exists, and its current value is saved in,
and restored from, EMUDESK.INF.  If CONF_WITH_BLITTER is 0, or the
machine does not have a blitter, the item will be greyed out.
In the future, clicking on the item will cause a check mark to be
added to indicate that the blitter is being used.  However, since
there is no actual blitter code present yet, the code to set/reset
the check mark is currently commented out.