Fix endmask bug in C version of bit_blt()
When the destination width was one word, endmask1 was always
calculated as the left endmask ANDed with the right endmask.  This
is incorrect when the x/y increments are negative (due to the source
address being less than the destination address): endmask1 should be
the right endmask ANDed with the left endmask.
Thanks to Christian Zietz for finding this problem!