Avoid including do_blit() unnecessarily
The do_blit() function was included when CONF_WITH_BLITTER and
ASM_BLIT_IS_AVAILABLE were both true, even though it could never be
called in this situation.  The #ifdefs have been reorganised to fix
this, saving approx 450 bytes in the 192K ROMs.
Thanks to Thorsten Otto for pointing this out.