Remove icons.c from POTFILES.in
POTFILES.in is a list of files that 'bug' scans for text to be
translated.  The file has been removed since there is no longer any
translatable text in icons.c.
Note: icons.c has been added as a dependency to makefiles.dep to
ensure that it is built at the start of the make process.  This was
previously driven by its place in POTFILES.in.