Change BITBLK colour for mouse cursors to black
The mouse cursor colours used by EmuTOS for display are taken from
the MFORM structure that is encapsulated within the BITBLK image;
these colours are normally black & white.  The value in the colour
field of the BITBLK structure is irrelevant for this; however,
resource editors will show the cursor in the BITBLK colour, so this
update changes that colour to black for consistency.