Do not use icon mask for desktop object selection
Previously, if the desktop view was by icon, clicking on an icon only
selected it if the click was on a pixel corresponding to a bit that
was set in the icon mask.  This is different from Atari TOS, where it
is only necessary to click on a pixel within the 16x16 icon area.
This update changes EmuTOS behaviour to correspond to Atari TOS.
Thanks to Christian Quante for the suggested change.