Fix alignment of desktop icons on a grid
If the icons did not fit exactly into the desktop work area, the
extra pixels were distributed only over the first half (left or top)
of the grid.  On large screens (e.g. TT mono), this resulted in
overlapping icon positions near the middle of the screen.
Many thanks to Christian Quante who reported the bug and also
provided the fix that was the basis for this patch.