Fix Mega STe boot problem
If any key was held down during EmuTOS boot, the keyboard and mouse
did not work.  This happened on a Mega STe, but was not reproducible
on a 1040ST or 1040STe.  The problem happened because the ACIA
signalled an interrupt before the MFP was programmed to detect one.
Fixed by reading the ACIA data register (to clear any pending
interrupt) immediately before configuring the MFP.