Add configurable retry for hard disk detection
Hard disk detection is performed via a call to internal_inquire().
This was previously done only once per device on configured/available
busses.  This patch allows the user to configure one or more retries.
Note that enabling retries will slow down the boot process, so the
number of retries is set to zero by default.