Set blitter state *after* reading EMUDESK.INF
Previously, the blitter state was set before EMUDESK.INF had been
read.  This caused it to be always off after a cold boot, whatever
the saved state of the blitter.  Thus the Blitter item could be
checkmarked, even though the Blitter was not enabled.  This
situation was corrected the next time EmuDesk was reloaded (e.g.
after running the first program).