Disable blitter on machines with TT-RAM
Because the Atari blitter only supports 24-bit addresses, we must
disable it if TT-RAM is present, since the latter requires 32-bit
addressing.  If it remained enabled, a user program could attempt to
use it to transfer to/from TT-RAM, and this would access the wrong
area of memory.  No Atari system was built with the blitter and
TT-RAM, but add-ons such as the CT60 have this combination.

Note that we do not restrict the FireBee, since it has a 32-bit
blitter.  Also, add-ons that add alternate memory below the 16MB
line are not affected.