Add -L (line number) option to 'bug'
Previous versions of bug included the source module line number in
the cross-reference in messages.pot.  This caused useless changes to
many lines in the .po files produced by 'bug update', when objects
or entire dialogs were inserted or deleted.  This made it harder to
monitor real changes to the text itself.
After this update, by default, line numbers are not included in the
cross-reference.  You may restore the previous behaviour by using
the -L option.