Fix EmuDesk mouse cursor initialisation
At the end of its initialisation, EmuDesk changes the mouse cursor
shape to an arrow.  It was possible for the control manager (which
manages the menu bar) to be called while this change was being made.
This could cause the mouse cursor shape to be left empty, or set to
an hourglass.  EmuDesk now grabs the screen drawing semaphore before
it changes the mouse cursor shape.  This bug was in the original
imported sources.