Fix bug in EmuDesk change resolution handling
If you double-click on an EmuDesk drive icon and then, before the
contents are displayed, press control-R, the "Change resolution"
dialog appears in front of a partially-built window.  If you select
Yes in the dialog, the resolution change does not take place and the
refresh of the partially-built window completes normally.  This
problem occurs when the evnt_multi() in the main desktop loop
returns MU_KEYBD and MU_MESAG at the same time: 'done' is set by
hndl_kbd() and then reset by hndl_msg().
Thanks to Vincent Rivière for finding this bug; the buggy code is in
the original imported sources.