Improve BDOS sector caching algorithm
BDOS cache flushing occurs in three places: usrio(), ixclose(), and
getbcb().  Previously, BDOS always invalidated a cache entry when it
was flushed, which caused unnecessary re-reads for commonly-accessed
sectors, such as directory or FAT sectors.  Now, cache entries are no
longer invalidated in ixclose(), which allows the cache to be used
more effectively.  The flush in usrio() and getbcb() still causes
invalidation, since these are poor candidates for longterm caching.