Avoid unnecessary directory sector writes in BDOS
When a directory is modified (e.g. by creating a file within it),
the OFD corresponding to the directory is marked as dirty.  When
ixclose() is subsequently called for that directory, the
corresponding sector in the cache is marked as dirty & is flushed.
However, the dirty flag in the OFD was never cleared, so the
directory sector in question was written to the disk *every time*
ixclose() was called for that directory, until the system was
restarted, even with no change to the sector.  This update reduces
the number of I/Os by up to 15% when copying small files.