Fix label bug when formatting floppy
The label field in the format dialog was not initialised.  This
caused two problems: (1) if the label field was never modified, a
string beginning with @ was written as a label; (2) if the label was
modified, the new value was retained in the prompt for the next
format.  Thanks to Christian Zietz for reporting this.