Fix design bug in VDI workstation creation
Opening a VDI workstation allocates memory which is owned by the
process which did the open.  However, in the original TOS/EmuTOS
design, the memory is chained on a global chain.  If the process
terminates without closing the workstation, the memory is freed,
but the freed memory remains on the chain.  This can/will lead to
crashes in subsequently-run programs.  We now maintain the pointer
to the workstation memory in system memory.