Fix retry count when switching floppy density
On systems with an HD floppy drive, if an HD diskette is replaced by
a DD diskette (or vice versa), the next I/O will fail with an error,
the density will be switched, and the I/O will be retried.  When this
happened, the retry count was not being reset to zero, so if this I/O
then encountered a temporary error, the I/O failed.  We now reset the
retry count in this situation.
Thanks to Christian Zietz for reporting the problem and suggesting a
solution.