Fix bug in floppy status decoding routine
The write protect bit in the floppy status returned by the FDC is
only valid after a write command.  The status decoding routine
always tested the bit, even for read commands.  This caused an
error to be detected when reading a floppy under Aranym.
The bug was introduced by commit c9e1a5e.