Initialize clockvec even when !CONF_WITH_IKBD_CLOCK
If !CONF_WITH_IKBD_CLOCK, clockvec is initialized with just_rts.
This avoids crashing if a clock paquet is received, in any case,
even with !CONF_WITH_IKBD_CLOCK.

This is particularly useful with CONF_SERIAL_IKBD when bogus data
is received on the serial port due to bad hardware, wrong baudrate, etc.