Fix system freeze when running FastCopy III
FastCopy III uses an undocumented feature of TOS's floppy-only
Rwabs() code: when the buffer pointer is NULL, the 'count' field is
used to set the mediachange value (0-2) for that drive, and no I/O
is actually done.  Previously, EmuTOS did not recognise this and
attempted to do a DMA read into location 0 (ROM) which resulted in
a system freeze on a real machine, a double bus error on Hatari.
Many thanks to Christian Zietz for finding this.