Fix filename display while loading a program
While loading a program from the desktop, Atari TOS displays just the
filename at the top of the screen, while EmuTOS displays the full
pathname.  For long paths, this is ugly (and may be truncated).
This update changes EmuTOS to act in the same way as Atari TOS.
Thanks to Vincent Riviere for pointing out this problem.