Fix drive B access under standalone EmuCON
By design, EmuCON does not support mapping of drive B to drive A on
a single-floppy system.  Previously, when running the standalone
version of EmuCON under Atari TOS on a single-floppy system, accesses
to drive B were mapped to drive A.  This caused confusion and could
cause problems for e.g. copy/move.  Now, in this scenario, drive B
is treated the same way as any other invalid drive.
Thanks to Vincent Riviere for identifying the problem.