Work around NVDI 5.03 incompatibility
When running in a 256-colour resolution, NVDI 5.03 tries to use the
same area as Atari TOS to save the mouse cursor background.  This is
beyond the end of documented lineA variables, causing a crash under
EmuTOS.  This commit extends the EmuTOS lineA variables area to fix
the problem.
Many thanks to Christian Zietz for his analysis of the problem.