Fix error message when exiting ORCS
If a program is run from the ORCS desktop, and then ORCS is exited,
EmuTOS issues an alert message ("The application cannot find the file
or folder you tried to access") corresponding to the AL18ERR error
code.  There should be no error message at this point.
Thanks to Thorsten Otto for reporting this.