Allow programs to change video mode via v_opnwk()
Programs running before the AES is initialized (e.g. from the AUTO
folder) are allowed to open the physical workstation, and can request
the video mode while doing so.  Previously, EmuTOS ignored this.
This change allows the boot manager XBOOT to boot properly on STs
which boot in low resolution mode.