Fix bug attempting to delete desktop file/folder
If a file or folder icon on the desktop is selected for deletion, an
initial prompt asks whether to delete the icon or the file.  If you
select file, nothing happens.  This fix makes the deletion behave
in the same way as if you deleted a file/folder in a window.