Ensure that EmuTOS only includes basic desk icons
The new desk icon resource introduced by commit b2bb762 had the
"extra" icons (the ones that should be excluded from the EmuTOS
image) in a tree called "OTHER".  However, the ird tool was looking
for the name "APPS".  This caused all of the icons to be included,
increasing ROM usage by about 2KB.  ird now looks for "OTHER".