Identify missing installed application
If an installed application is no longer available (for example,
because it has been moved or renamed), then documents of the relevant
type will invoke a non-existent program.  Previously, this would
cause a general AES alert to be issued which did not identify the
offending program.  Now, in this situation, EmuDesk issues a form
alert showing the name of the missing program.