Change 'check for parent' test for copy/move
It is invalid to copy/move a parent folder into a child folder (since
this leads to an infinite recursion). Previously, EmuDesk's test to
prevent this was made before the copy/move was started, but the test
was invalid if filemasks were in use.  It could also stop otherwise
valid copies/moves, made at the same time as the recursive copy/move,
from proceeding.  The check is now made during the actual copy/move
process, just like Atari TOS does.