Remove useless test for nesting in 'New Folder'
The old code that implemented the 'New Folder...' EmuDesk menu item
did not allow the creation of a folder which would be more than 8
levels deep from the root.  This is the same as Atari TOS 1.
However, this restriction was removed in TOS2/3/4, since it serves
no useful purpose, and now it is removed in EmuTOS.
Note that EmuTOS (and Atari TOS) still restrict the level of
recursion allowed for copy/move/delete, but this is relative to the
current top of the directory being acted on, not the root.