Allow space in shell buffer for accessory data
Old versions of the control panel and other accessories expect that
the shell buffer will start with an area for their use, beginning
with '#a' and 128 bytes long in total.  This update provides that
area, but does not initialise it.  By design, unlike Atari TOS, the
area is NOT saved to the EMUDESK.INF file.