Fix bug in EVB handling
If there are no EVBs on the EVB unused list (eul), iasync() will
crash trying to access location 0.  This commit does two things:
increases the number of EVBs to the maximum that can be used; and
issues a panic() if the "impossible" happens.  Thanks to Christian
Zietz for pointing out this problem.