Fix bug in code that merges redraw messages
The existing code had a test for a value of 0xFFFF in the "additional
message length" word of system messages.  The test was incorrectly
formed, but in any case unneeded since system messages always have a
value of 0 in that word.  This bug was in the original DR imported
sources.  Thanks to Eero Tamminen for reporting this.