tests: Check correct values in the cookie jar
We can use the value in the cookie jar to check whether EmuTOS detected
the machine type and CPU type correctly. This patch adds a little TOS
program that dumps the cookie jar. We then use Hatari to run EmuTOS
with this little test program with various machine and CPU types and
then analyze the dumped cookie jar values to see whether they match
the emulated hardware.