tests: Add a serial loopback test with Bconin/Bconout
Check whether Bconin(1) and Bconout(1) are working correctly by
running Hatari with a file that contains random data as input for
the RS232 interface. The TOS test program then writes this input
data via Bconout(1, ...) back to the RS232 output. At the end
of the test, we compare the input with the output file to make
sure that everything was handled correctly.