Display hourglass while launching program
When launching a program from the desktop, the mouse cursor should
remain as an hourglass until the program itself sets it to an arrow.
However, since the desktop exits via appl_exit() before the new
program is run and the old version of appl_exit() switched to the
arrow cursor, the mouse cursor switched to the arrow too soon.