Circumvent crash in Chameleon DA
The Chameleon accessory intercepts a number of system vectors
including the BIOS and BDOS traps.  A resolution change caused a
crash, because Chameleon relies on seeing an explicit Mfree() of
its own memory to trigger a cleanup.  This did not happen because
EmuTOS freed the memory directly rather than via Mfree().  Fixed
by explicitly freeing DA memory.
Thanks to Eero Tamminen for reporting the problem.