Fix screen display on Falcon with NVDI 4
This fixes a problem due to a discrepancy between the hardware screen
settings and NVDI's software settings.  The problem may also apply to
other versions of NVDI.  The symptoms are a messed-up screen repeated
several times across the real screen.  Thanks to Christian Zietz &
Thorsten Otto for their help in identifying the root problem.