Add Clr/Home and Insert bits to 'shifty'
Atari TOS sets flag bits in the 'shifty' byte if either the Clr/Home
or the Insert key is pressed while the Alt key is down.  We now do
the same.  Note that this is poorly documented in the usual sources
(the Compendium, the Hitchhiker's Guide to the BIOS), and the EmuTOS
implementation is based on tests using Atari TOS 2.06/3.06.