Switch travis builds from 'xenial' back to 'trusty'
We're currently facing some issues with the deployment step... so switch
back to 'trusty' to sort out that the problem has been caused by the update
to 'xenial' recently.
With the older version of Hatari in 'trusty', the bioscon test does not
finish in time, so the amount of maximum VBLs has to be increased for this
test now.