Revert "Fail immediately if makefile.dep is missing"
This reverts commit 90155048ba01c0017d88553376e4cc74bb0aaed8.

Scenario:
make clean
make 512

GNU make 3.81 prints a warning:
Makefile:1285: makefile.dep: No such file or directory

It seems that make 3.81 always complains if an included makefile is missing,
even if it is going to build it by itself through the auto-rebuild feature.

So we go back to the previous strategy:
Include makefile.dep using -include
The initial hyphen asks include to silent the warning if the file is missing.