Prevent useless alerts when starting EmuCON
If EmuCON is started on a system with no media (e.g. floppy-only with
no disks loadaed), two form alerts are issued due to AES trying to
determine the current path.  This is saved in a global that has no
useful purpose.  The global is now removed along with the calls that
generated the alerts.