Improve conditional compilation for bootstrap()
Currently, bootstrap() is only available through ARAnyM's NatFeats.
With this mechanism, the emulator can provide an external PRG
just as if it was the first in the AUTO folder.
This is useful to load MINT.PRG, when the driver for the boot partition
(FAT32, ext2, hostfs) is not provided by EmuTOS.