Tweak window size change after resolution change
If the saved desktop window coordinates (xywh) cause a window to
be higher or wider than the desktop when opened in the current
resolution, then the window width and/or height is automatically
modified to fit the desktop.  The previous algorithm usually made
the window too small and/or narrow.  Note that this change applies
both for windows that are open when the resolution is changed, and
when loading an EMUDESK.INF that was saved at a higher resolution.