Rework detection of IDE interfaces incl. twisted cables and ghosts
Previously, the code would first try to detect an interface with
a regular cable, and, if that timed out, an interface with a twisted
cable. Particularly when multiple IDE interfaces were detected,
possibly due to incomplete address decoding ("ghost interfaces"),
multiple 31s timeouts occured. This lead to the impression that
EmuTOS was hanging during boot.

Now, regular and twisted cables are checked in parallel to reduce
the number of timeouts. Thanks to Marcelo Pires for reporting this.