Increase text scratch buffer size
If text is output with a style including OUTLINE and SKEW, a non-zero
rotation, and no scaling, the rotation of the outlined text takes
place in the small part of the buffer.  This is too small, and the
data overflows into the large part of the buffer.  The overflow is
apparently harmless since it occurs after that part of the large
buffer has been processed.