Display desktop function key assignment
This allows the user to display the function keys (and corresponding
programs) that have been assigned via the "Install application..."
menu item.  To display, select the "Desktop configuration..." menu
item and use the arrows: up/down arrows select the function keys,
and left/right arrows scroll along the full pathname of the program.
This is modelled after the same feature in Atari TOS 2 and later.