Fix bug in bit_blt()
The planes of the foreground & background colours specified were
processed in reverse order.  For example, a hardware colour of 3
(=0011b) was treated as a colour of 12 (=1100b).  This could cause
bad output for vrt_cpyfm(), linea7(), and lineae() when bit_blt()
was used.  bit_blt() is used when the blitter is enabled, or when
EmuTOS runs on Coldfire [otherwise fast_bit_blt() is used].
This bug was introduced by the initial conversion from assembler
to C, in mid-2004.