Revert "Add missing space before KB"
Notes from Roger:
The length of space provided (underscores) is 19.  The maximum possible string
is something like: "14000KB + 1048576KB" which is 19 chars long.  Putting the
extra spaces in will potentially cause the string to overflow.  And increasing
the number of underscores will cause the object to overflow the dialog box when
the translation is long (see the Russian dialog box).

This reverts commit 99934ed1f74553800cf4212f8ad3563c8857f8a3.