Fix VDI line clipping calculation
For line clipping, the VDI uses clip_line(), which uses mul_div().
In the original DRI code, the equivalent assembler function
SMUL_DIV() performed rounding before returning the result.  When
this was implemented in C after the original import of sources,
the rounding was dropped.  The current assembler re-implementation
also omits the rounding.  This caused bad output for some sloped
lines when clipping was in effect.

Rounding has now been restored in a new assembler function called
mul_div_round().  Initially, this is only called in clip_line().