Fix filled polygon drawing properly
The routine to fill a polygon is called directly or indirectly by
many functions, including arrow(), wideline(), v_fillarea(), and
many of the GDP functions.  It calls clc_flit() to determine the
start and end points of each line to be drawn.  The existing
clc_flit() excluded the perimeter points when called from the VDI.

However, testing under Atari TOS shows that the fill area always
*includes* the left and right perimeter points (for those functions
that allow the perimeter to be drawn separately, it is drawn on top
of the edge pixels).  We now conform to Atari TOS.

This change refixes the problem fixed by commit 4e52a6a and also
fixes a number of problems with the functions listed above.