Refix obscure 'blinking cursor' bug
If the last thing a program wrote to the console (before terminating)
was an Esc character, the 'disable alpha cursor' VT52 escape sequence
sent indirectly by sh_tographic() was ignored.  This caused a
blinking cursor to be left on the desktop.  This bug was originally
fixed by commit d3e3480, but that fix caused a crash when switching
resolutions with fVDI installed.  Thanks to Christian Quante for his
report, and to Christian Zietz for his debugging.