Improve VDI text output speed for 'plain' text
This increases VDI text output speed by a factor of 4 for a limited
subset of output.  The most significant restriction is that the
output must be byte-aligned on the screen.  Other restrictions
include: no effects, no rotation, text must be left-aligned.  Also,
the font must be monospaced with a width of 8 pixels, and the output
string must not be clipped.  The equivalent feature exists in all
versions of Atari TOS.