Implement 'Size to fit' desktop menu item
This does the same as the equivalent Atari TOS 2/3/4 menu item: it
provides control of the window display.  If 'Size to fit' is enabled,
the width of a window controls the number of columns of files that
are displayed, and the horizontal scroll bar and arrows are unused.
If 'Size to fit' is disabled,  a desktop window is treated as a view
onto a virtual window that is the width of the desktop, and both
vertical and horizontal scrolling are possible.
By default, EmuDesk windows are always 'Size to fit'; for code space
reasons, this may not be changed in the 192K ROMs.