Fix compatibility with Geneva 1.8
This bug has been introduced in commit e1455dee.
Thanks to Mia Jaap for having spotted this bug,
and to Christian Zietz for the investigation.

Without this, Geneva fails with:
"Something is already hooked into exec_os"

Solution is to move ui_mupb at the end of the ROM.

Note: Older Geneva Release 003 still crashes for other reasons.