Fix bug in lineA undraw_sprite()/draw_sprite()
It is legal to obtain the address of the lineA routines via linea0,
and then call them directly from supervisor state.  The existing
code did not handle this correctly for undraw_sprite()/draw_sprite().
This caused a crash in "Art for Kids".  Many thanks to Eero Tamminen
for reporting the problem.