Clear caches during 020/030/060 processor init
During initialisation, the 68040/68060 instruction/data caches are
already cleared.  We now clear the 68020/68030 instruction/data
caches and the 68060 branch cache as well.  This is not required when
EmuTOS runs after a reset (e.g. when running from ROM), since the
reset clears the caches.  However, it may be required when starting
EmuTOS as a program.