Implement 'Cache' desktop menu item
This does the same as the equivalent Atari TOS 3/4 menu item: it
allows the CPU cache(s) to be enabled or disabled.  By design, this
only applies to 680x0 CPUs and is currently available in all of the
relevant EmuTOS versions except the 192K ROMs.
If the current CPU does not support caching (68000/68010), the 'Cache'
menu item is greyed out.  Otherwise, by default, it is checked and the
caches are enabled.  Unchecking the menu item disables the caches.
The cache status is maintained in EMUDESK.INF.