Prompt user when an orphan shortcut is dragged
If a desktop shortcut had no corresponding file (due to a file move
or delete), dragging the shortcut to a window or another icon (e.g.
trash) was silently ignored.  We now prompt the user with the
'Remove icon or locate file' dialog.  Thanks to Dima Sobolev for
pointing this out.