Fix memory overwrite by file selector
The file mask string obtained from the input path parameter was
copied to the string pointed to by the te_ptext field of the file
selector's title line.  Following commit 390f282 (Dec/2014), this
field is only 1 byte long, and therefore the mask usually wrote
past the end of the string.  By chance, this usually did not cause
noticeable problems.